refactor: move test files to separate folders on server
